% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawr.R
\name{PaginateAggs}
\alias{PaginateAggs}
\title{Paginate aggs}
\usage{
PaginateAggs(type = c("comment", "submission", "subreddit"),
  aggs = c("author", "link_id", "created_utc", "subreddit"),
  paginate_by = c("date", "author"),
  before = round(as.numeric(Sys.time())), after = NULL,
  timescope = NULL, verbose = getOption("PAWR.VerbosePaginate"), ...)
}
\arguments{
\item{type}{Type of requested content. Can be \code{comment}, \code{submission}, or \code{subreddit}.}

\item{aggs}{What should be aggregated over?}

\item{paginate_by}{Define which variable should be used to break the data into smaller chunks; either \code{author} or \code{date}.}

\item{verbose}{Should output be verbose? Defaults to a global option which can be set with \code{options(PAWR.VerbosePaginate=TRUE/FALSE)}.}

\item{...}{Other valid parameters. Run \code{PSParams()} to see all valid parameters and their descriptions.}
}
\description{
Send multiple queries to pushshift.io to get all available information for your request.
This function is meant to get around the maximum of 1000 items returned by a single aggs query.
}
\examples{
#Find out on which subreddits the users of r/cheese post
#Analysis is limited to December 2019
users<-PaginateAggs(aggs="author",paginate_by="date",
  subreddit="cheese",timescope=30*24*60*60,before=1577836800)
users<-users$key
#remove bots and missing values
users<-users[!(users \%in\% c("[deleted]","AutoModerator"))]
#Posting behavior of all authors is aggregated.
subreddits<-PaginateAggs(aggs="subreddit",paginate_by="author",
  author=users,timescope=30*24*60*60,before=1577836800)
}
