% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawr.R
\name{PaginateData}
\alias{PaginateData}
\title{Paginate PushShift Data}
\usage{
PaginateData(type = c("comment", "submission", "subreddit"),
  verbose = getOption("PAWR.VerbosePaginate"),
  before = round(as.numeric(Sys.time())), after = NULL,
  timescope = NULL, ...)
}
\arguments{
\item{type}{Type of requested content. Can be \code{comment}, \code{submission}, or \code{subreddit}.}

\item{verbose}{Should output be verbose? Defaults to a global option which can be set with \code{options(PAWR.VerbosePaginate=TRUE/FALSE)}.}

\item{before}{Upper limit in the date range of posts to be fetched.}

\item{after}{Lower limit in the date range of posts to be fetched.}

\item{timescope}{Time range, in seconds, within which posts should be fetched;
works in conjunction with either \code{before} or \code{after}.
When argument \code{after} is used, \code{timescope} causes the current function to fetch data that was created up to \emph{N} seconds after the timestamp defined in \code{after}
When argument \code{before} is used, \code{timescope} causes the current function to fetch data that was created up to \emph{N} seconds before the timestamp defined in \code{before}}

\item{...}{Other valid parameters. Run \code{PSParams()} to see all valid parameters and their descriptions.}
}
\description{
Send multiple queries to pushshift.io, to be able to get all data within a given date range.
}
\examples{
#Get all comments from today containing the word "chocolate"
PaginateData(timescope=24 * 60 * 60,q="chocolate")
}
