% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawr.R
\name{QueryPushshift}
\alias{QueryPushshift}
\title{Query data from pushshift.io}
\usage{
QueryPushshift(
  type = c("comment", "submission", "subreddit"),
  as.df = T,
  purge = F,
  verbose = getOption("PAWR.VerboseGet"),
  size = getOption("PAWR.QuerySize"),
  aggs = c("none", "author", "link_id", "created_utc", "subreddit"),
  agg_size = 0,
  q = NULL,
  metadata = TRUE,
  ...
)
}
\arguments{
\item{type}{Type of requested content. Can be \code{comment}, \code{submission}, or \code{subreddit}.}

\item{as.df}{Convert output to \code{data.frame}? Defaults to \code{TRUE}.}

\item{purge}{Purge deleted posts? Defaults to \code{FALSE}.}

\item{verbose}{Should output be verbose? Defaults to a global option which can be set with \code{options(PAWR.VerboseGet=TRUE/FALSE)}.}

\item{size}{Maximum number of pieces of content to return. Defaults to the maximum, which is 500, except when using aggs, when it defaults to 0.}

\item{agg_size}{Maximum number of values to return when using aggs; defaults to 500 unless you're not using aggs.}

\item{q}{Query term.}

\item{metadata}{Request metadata from pushshift, which will be used to check whether the query was successful or some of pushshift's shards failed to respond. This is recommended for academic research.}

\item{...}{Other valid parameters. Run \code{PSParams()} to see all valid parameters and their descriptions.}
}
\value{
If \code{as.df=T}, returns a data.frame; else, returns a list.
}
\description{
Query data from pushshift.io
}
\examples{
#Get u/spez's first ever comment
QueryPushshift(author="spez",after=0,size=1)

#See in which subreddits the word "gamer" is used the most
GetPSData(q="Gamer",aggs="subreddit")
}
