% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pawr.R
\name{UtilityFunctions}
\alias{UtilityFunctions}
\alias{refreshPAWR}
\alias{list2df}
\alias{unevenrbind}
\alias{PSParams}
\alias{GetTotalQuerySize}
\title{Utility Functions}
\usage{
refreshPAWR(verbose = T)

list2df(li)

unevenrbind(df1, df2)

PSParams(type = c("all", "comment", "submission", "subreddit"))

GetTotalQuerySize(...)
}
\arguments{
\item{verbose}{Logical. Produce verbose output or not.}

\item{li}{List of lists, to be converted to \code{data.frame}.}

\item{df1, df2}{To-be-merged data frames with an uneven number of columns and/or nonmatching column names}

\item{type}{Character. The type of content that parameters are being looked up for (comment, submission, subreddit). Defaults to all.}

\item{...}{For GetTotalQuerySize, any parameters to be sent to pushshift.io}
}
\description{
Utility Functions
}
\details{
\code{refreshPAWR()} re-fetches the rate limit and parameter list from pushshift.io
\code{list2df()} converts a list of lists (pushshift's preferred output format) to a data.frame
\code{unevenrbind()} binds two data.frames together by row, even if their columns do not match
\code{PSParams()} returns all available pushshift.io parameters for a given data type; defaults to all data types.
\code{GetTotalQuerySize()} gives the total amount of content that matches the provided parameters
}
